Refactored to use OutputWriter from opm-output
As OutputWriter has been deleted from opm-core
Only changes in includes
