+++ opm/autodiff/SimulatorFullyImplicitBlackoilOutput.cpp
@@ -22,7 +22,7 @@
 #include "SimulatorFullyImplicitBlackoilOutput.hpp"
 
 #include <opm/core/utility/DataMap.hpp>
-#include <opm/core/io/vtk/writeVtkData.hpp>
+#include <opm/output/vtk/writeVtkData.hpp>
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/core/utility/miscUtilities.hpp>
 #include <opm/core/utility/Units.hpp>
