+++ opm/autodiff/SimulatorFullyImplicitBlackoilSolvent.hpp
@@ -44,7 +44,7 @@
 #include <opm/core/simulator/SimulatorTimer.hpp>
 //#include <opm/core/simulator/AdaptiveSimulatorTimer.hpp>
 #include <opm/core/utility/StopWatch.hpp>
-#include <opm/core/io/vtk/writeVtkData.hpp>
+#include <opm/output/vtk/writeVtkData.hpp>
 #include <opm/core/utility/miscUtilities.hpp>
 #include <opm/core/utility/miscUtilitiesBlackoil.hpp>
